% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abmi.themes.R
\docType{package}
\name{abmi.themes}
\alias{abmi.themes}
\alias{abmi.themes-package}
\title{abmi.themes: Making ABMI-themed ggplot2 graphics}
\description{
This package provides ABMI staff with access to colour palettes and web fonts outlined in the Brand Guidelines, which can be used when making graphics wiht ggplot2. A default ABMI ggplot2 theme is also provided.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mabecker89/abmi.themes}
}

}
\author{
Marcus Becker \email{mabecker@ualberta.ca}
}
