% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{abmi_cols}
\alias{abmi_cols}
\title{Function to extract ABMI colours as hex codes}
\usage{
abmi_cols(...)
}
\arguments{
\item{...}{Character names of abmi_colours (case insensitive, partial matching allowed)}
}
\value{
A name vector of hex colors
}
\description{
Function to extract ABMI colours as hex codes
}
\examples{
\donttest{
abmi_cols("rhino", "pewter", "locust")
}
}
