% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{abmi_pal}
\alias{abmi_pal}
\title{Return function to interpolate an ABMI color palette}
\usage{
abmi_pal(palette = "main", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in abmi_palettes}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments to pass to colorRampPalette() (e.g. alpha value)}
}
\value{
A list of hex colours
}
\description{
Return function to interpolate an ABMI color palette
}
