% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_abmi_fonts.R
\name{add_abmi_fonts}
\alias{add_abmi_fonts}
\title{Load ABMI official fonts: Montserrat and Domine}
\usage{
add_abmi_fonts()
}
\description{
Load ABMI official fonts: Montserrat and Domine
}
\examples{
\dontrun{
add_abmi_fonts() # call before plotting
}

}
