% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_logo.R
\name{add_logo}
\alias{add_logo}
\title{Add ABMI logo to the bottom of the plot}
\usage{
add_logo(plot_name, logo = "acronym", logo_position = 0.925)
}
\arguments{
\item{plot_name}{The variable name of the plot you have created}

\item{logo}{Character; Specify which logo you want (see details below)}

\item{logo_position}{A numeric vector or unit object specifying x-location of logo; defaults to 0.925 (right justified)}
}
\value{
This function returns your plot with the ABMI logo at the bottom right
}
\description{
Add ABMI logo to the bottom of the plot
}
\details{
Valid values for the \code{logo} argument currently are:
\itemize{
\item "acronym" (default)
\item "full"
\item Personal file path for a logo image you want to use (needs to be a png file)
}
}
\examples{
\donttest{
add_logo(p, logo = "full", logo_position = 0.875)
}
}
