% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dep}
\alias{dep}
\title{ABMI deployments by year}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 5 rows and 2 columns.}
\source{
ABMI
}
\usage{
dep
}
\description{
ABMI deployments by year
}
\keyword{data}
