% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mammals}
\alias{mammals}
\title{Number of mammal images}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 8 rows and 2 columns.}
\source{
ABMI
}
\usage{
mammals
}
\description{
Number of mammal images
}
\keyword{data}
