% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{print_palette}
\alias{print_palette}
\title{Print palette}
\usage{
print_palette(x, ...)
}
\arguments{
\item{x}{The palette as a character vector of hex codes}

\item{...}{Arguments to pass to \code{image()}}
}
\value{
An image of the requested colour palette
}
\description{
Print palette
}
