% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{scale_color_abmi}
\alias{scale_color_abmi}
\title{Colour scale constructor for ABMI colours}
\usage{
scale_color_abmi(palette = "main", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in ABMI palettes}

\item{discrete}{Boolean indicating whether colour aesthetic is discrete or not}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments passed to \code{discrete_scale()} or \code{scale_color_gradient()}, which are used when discrete is TRUE or FALSE, respectively}
}
\description{
Colour scale constructor for ABMI colours
}
