% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_logo.R
\name{add_logo}
\alias{add_logo}
\title{Add ABMI logo to the bottom of the plot}
\usage{
add_logo(
  plot_name,
  logo_image_path = file.path(system.file("inst/extdata", package = "abmi.themes"),
    "horizontal-logo-acronym.png")
)
}
\arguments{
\item{plot_name}{The variable name of the plot you have created}

\item{logo_image_path}{File path for the logo image you want to use in the right hand side of your plot, which needs to be a PNG file; defaults to the ABMI's primary horizontal acronym logo that is stored in the data\ folder of this package}
}
\value{
This function returns your plot with the ABMI logo at the bottom right
}
\description{
Add ABMI logo to the bottom of the plot
}
\examples{
\donttest{
add_logo(p)
}
}
