% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_abmi.R
\name{theme_abmi}
\alias{theme_abmi}
\title{Add custom ABMI theme to a ggplot graphic}
\usage{
theme_abmi(font = "Montserrat")
}
\arguments{
\item{font}{Defaults to "montserrat"; alternative is "domine" (case insensitive, partial matching allowed)}
}
\description{
Add custom ABMI theme to a ggplot graphic
}
\examples{
\dontrun{
p + theme_abmi()
}
}
